Update ASDF::*DEFINED-SYSTEMS* after leaving local Quicklisp.
After leaving the local Quicklisp environment, update
ASDF::*DEFINED-SYSTEMS* to include the systems that were loaded from the
local environment. This keeps ASDF up to date with the systems that are
actually loaded and allows things like FIND-SYSTEM and
SYSTEM-RELATIVE-PATH NAME to work as expected for loaded systems without
needing to wrap in a WITH-LOCAL-QUICKLISP.

Also has the side effect of letting ASDF know of the switch from systems
loaded outside the environment to systems loaded inside the environment.