Allow git to inherit the lisp's stdin for terminal detection.
When cloning via ssh, the ssh process that git starts needs to
prompt for a password, which requires a terminal or an ssh-askpass
program. When the input is not a terminal (the default) but ssh is
being run in a terminal session, ssh detects the following:
1) stdin is not a terminal, so we can't prompt there, but
2) there /is/ a controlling terminal for the process, so we
shouldn't use ssh-askpass to produce a graphical prompt.

This causes ssh to hang, and makes ssh-based git sources unusable
when running e.g. QLOT:INSTALL from an actual terminal. Using
:INPUT :INTERACTIVE allows ssh to use the terminal in terminal
sessions, and to use ssh-askpass when run from a non-terminal
(e.g. emacs).