Preserve pathname components when getting a directory pathname.
Pathnames contain more components than just the directory, name, and type;
if you create a directory pathname out of a name, type, and directory,
you'll lose the other parts of the original pathname. By using :defaults,
you can un-set specific components and inherit the rest from the original
pathname.

This fixes qlot:quickload on Windows, which stores the drive letter of
paths in the DEVICE component of pathnames.