Apply hacky, partial mitigation for quicklisp client issue.
When installing a system, fetching a dist object by using
(QL-DIST:DIST "name") results in a dist object whose system-index is not
properly initialized. This will result in the PROVIDED-SYSTEMS for one of
the dist's releases including a bunch of entries that are NIL, because
FIND-SYSTEM-IN-DIST was unable to locate the system in the index. That
causes release installation to fail, because (QL-DIST:INSTALL NIL) is not
a valid call.

While I believe this to be a bug in quicklisp, qlot can mitigate the issue
by using the dist object returned by QL-DIST:INSTALL-DIST (which /does/
have a properly initialized system index) when installing releases.