Reduce code loaded by non-quicklisp sources during distification.
This removes an unused package dependency, and moves a utility
specific to the quicklisp source to a utility package only used by
that source.

This cuts down on the amount of code that needs to be loaded by the
distification sub-process for non-quicklisp sources; at present all
source types are loaded during distification, but this paves the way
for future performance improvements by only loading the code necessary
for a single source type during distification.