Use a common utility function to write standard metadata.
A number of sources need to write dist metadata that is generated in
the same way; there are already utility functions for doing each part
of it, but sources are still duplicating code to perform the same
process. Using a common utility cuts down on duplication, shortens the
code a little, and centralizes information about the metadata file
names to be generated.

In addition to a modest reduction in duplication, this is also
important for future work that will reduce the scope of the
distification sub-process: this utility function is the only part of
distification with side-effects, so it's the only bit that needs to be
run in a sub-process. By having a single function available in a
lightweight package with fewer dependencies -- not the source type's
distify package, which pulls in a big chunk of the source packages --
distification can be faster because there is less code to load.