Overhaul distification, version-lock sources in the host.
This is a major overhaul of the distification process. The
highest-level goal here is to have sources be version-locked in the
host, rather than a sub-process, so that they will have locking
information when they are dumped to the lockfile.

To do this, distification has been divided into a three-stage process:
1. Prepare the source, e.g. by downloading a tarball
2. Lock the source version
3. Distify the locked source

Preparation, version-locking, and most of distification can be run in
the host; the only part that requires a sub-process is performing
system discovery at the last stage of distification (i.e. writing
systems.txt). SOURCE-HTTP is the only source type that truly needs a
prepare step; most of the others can do version-locking first, and use
the locked version to do a download as part of distification. These
steps are implemented as a generic-function protocol implemented by
each source type, with QLOT/DISTIFY:DISTIFY driving the whole process.

By pushing the sub-process all the way down into the system discovery
code, the sub-process can load a much smaller module (just the
distification code for the source type being distified), which makes
distification faster. A cold install of a simple three-source system
showed a 5 second speed improvement with a hot system quicklisp
cache. The speedup would be larger with more sources or a cold system
quicklisp.

The patch here is large, but here's a high-level summary of notable
changes:
  - QLOT/DISTIFY-PROTOCOL was introduced with the distification
    protocol functions and a stripped-down generic replacement for
    RUN-DISTIFY-SOURCE-PROCESS.

  - The sources' distification code was converted to implement the
    generic functions in the protocol; full distification is typically
    separated from the DISTIFY-SOURCE method definition, because
    combining them obscures the early-exit behavior when
    :DISTINFO-ONLY T is passed. These run the final
    WRITE-STANDARD-METADATA in a sub-lisp where necessary.

  - QLOT/SERVER:*SYSTEM-QUICKLISP-HOME* was exported, and moved to
    QLOT/UTILS/QL for dependency reasons. This is still bound in
    QLOT/SERVER:WITH-QLOT-SERVER, but the definition is next to
    WITH-QUICKLISP-HOME, which seemed appropriate.

  - The portions of the code that used to call
    RUN-DISTIFY-SOURCE-PROCESS now call QLOT/DISTIFY:DISTIFY intead;
    this now runs in the host, not in a
    sub-process. RUN-DISTIFY-SOURCE-PROCESS is defunct and has been
    removed.

  - The unit tests have been updated to use the top-level DISTIFY
    function instead of the source-specific DISTIFY-<source type>
    functions used previously. Since the distification tests no longer
    run "inside" the sub-process boundary, those tests are also
    wrapped with WITH-QLOT-SERVER to set up the appropriate bindings.