Fix github source test data.
The lockfile in the test data has a github source with a version that
does not correspond to any ref in the upstream repository. The version
is instead the md5sum of the archive file, as expected by an http
source, rather than a github one.

This was exposed by fixing the previous issue with incorrect github
source URLs. Now that qlot is attempting to download the correct
archive, it's apparent that the github source is using an archive md5 as
the source version instead of a git ref, as if it were an http source.

This updates the test data to use a git ref, and not one from the tip
of the repo, as a way to exercise version-locked source installation.