Use version-locked ref for github source urls.
Previously the code constructed archive urls out of the ref explicitly
set in the source definition, but did not take the locked source
version into account. For sources that had been locked to a version,
the url would be incorrect if the source definition did not specify an
identical ref in its arguments.

This also adds a small accessor for indicating whether a source has
been version-locked (i.e. it's safe to call SOURCE-VERSION), since
SOURCE-VERSION refers to a slot that is unbound by default.

Note that the tests still fail, since this has turned up a new bug.