Fix quicklisp dist source target dir.
Sources can only be distified into/installed from the temporary directory
created by WITH-QLOT-SERVER, because it sets up the quicklisp fetch
handlers to look for files there. This used to work because the
duplicate DISTIFY call inside WITH-QLOT-SERVER would create a distinfo
file in the right place, even though this code was later re-creating it
in the qlhome as part of a fully DISTIFY. This is no longer the case
now that distification has been de-duped.

Using qlhome as a directory to generate a dist in also doesn't make a
lot of sense, so this was probably unintentional to begin with.