Fix reading of initial tarball entry, add note about tarball reqs.
READ-ENTRY-FROM-ARCHIVE reads only an entry header from an archive
file, and DISCARD-ENTRY or EXTRACT-ENTRY must be called afterwards to
read past the file data so that the stream is correctly positioned for
the next read. Do that with the initial entry we read by hand, so we
don't screw up subsequent reads.

The previous code here sort of worked, since we expect the first entry
to be a directory, which doesn't have any file data. However, this
caused the extraction function to choke on other tarballs, but not in
a way that really explained what was going on.

I've also added two notes about the required structure of tarballs to
be extracted this way, since that wasn't apparent to me
initially. There are a whole bunch of ways that this function, and
tarball support in general, can fail if the tarballs aren't exactly
the structure we expect, but that's a project for another time.