Eliminate unnecessary nesting of WITH-QUICKLISP-HOME.
In the original version, qlhome was set to the project-local qlhome,
only to have it be immediately reset to the system-qlhome around every
use of WITH-QLOT-SERVER, which would immediately set it back to the
project-local qlhome before executing its body. This is correct, but
confusing and hard to follow because it results in triply- or
quadruply- nested changes to qlhome, on top of what the code that gets
called does.

Instead, only use the project-local qlhome around the dist lookup, and
allow WITH-QLOT-SERVER to set it for the code that needs it. This
eliminates the need to switch back to the system qlhome around
WITH-QLOT-SERVER, which means we no longer have to save the system
qlhome before the loop.

This is mostly just a straight hoist of the inner code; the uninstall
step for updating the "quicklisp" dist had to be pushed down into
WITH-QLOT-SERVER as well. This also consolidates the loop's
WITH-PACKAGE-FUNCTIONS forms to cut down on the number of extra
wrapper forms.